+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-- add `medic run` for executing arbitrary shell commands with consistent
+- Allow shell commands to run with `inline = true` to disable progress bars
+  and inherit stdio directly from the parent medic process.
+- Add `medic run` for executing arbitrary shell commands with consistent
   formatting, progress spinners.
 - Use spinners for progress indicators for checks and steps.